+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Commands/ExitCommand.cs
@@ -1,4 +1,5 @@
 using System.Windows;
+using TD.AppInfra.Commands;
 
 namespace TD.WPF.Commands;