+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Commands/Maintenance/ExchangeDeleteCommand.cs
@@ -1,4 +1,5 @@
 using System.Windows;
+using TD.AppInfra.Commands;
 using TD.i18n;
 using TD.WPF.ViewModels;