+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Commands/Maintenance/SymbolCreateCommand.cs
@@ -1,4 +1,5 @@
 ﻿using System.Windows;
+using TD.AppInfra.Commands;
 using TD.AppInfra.Services;
 using TD.AppInfra.ViewModels;
 using TD.i18n;