+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Commands/Maintenance/TradingAccountCreateCommand.cs
@@ -1,4 +1,5 @@
 using System.Windows;
+using TD.AppInfra.Commands;
 using TD.AppInfra.Services;
 using TD.i18n;
 using TD.Models;