+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Commands/ShowAboutCommand.cs
@@ -1,4 +1,5 @@
 using System.Windows;
+using TD.AppInfra.Commands;
 using TD.WPF.Dialogs;
 
 namespace TD.WPF.Commands;